// app/src/shared/api/exchange/types.ts
export type City = SecondaryPropertiesOption;
export type Bank = SecondaryPropertiesOption;
export type Network = SecondaryPropertiesOption;
export type Currency = {
  id: number;
  icon: string;
  name: string;
  cities?: City[];
  banks?: Bank[];
  networks?: Network[];
};

export type GetCurrenciesGetApiResponse = Currency[];
export type GetCurrenciesGetApiArg = {
  /** Идентификатор отдаваемой валюты */
  giveCurrencyId?: number | null;
  /** Тип валюты ('BANK' — банковская валюта, 'COIN' — криптовалюта, 'CASH' — наличная) */
  currencyType?: string | null;
};

type SecondaryPropertiesOption = {
  name: string;
  /** ID сети */
  id: number;
};

export type GetDirectionInitialDataByDirectionTypeApiResponse =
  /** status 200 Успешный ответ */ {
    rate?: {
      /** ID направления обмена */
      id: number;
      /** Текущий курс обмена */
      course: number;
      /** Формат отображаемого курса */
      course_view: string;
      /** Минимальная сумма для обмена */
      currency_give_min_value: number;
      network: Network;
      city: City;
      bank: Bank;
      /** Тип направления обмена */
      direction_type: DirectionType;
      currency_give: Currency;
      currency_get: Currency;
      course_title: string;
    };
    currencies_give?: Currency[];
    currencies_get?: Currency[];
  };
// ...
export type DirectionType =
  | "COIN - BANK"
  | "COIN - CASH"
  | "BANK - COIN"
  | "CASH - COIN";